refactor into a manager
